feat(marketing): move to scope marketing pages
